p2p: make nTime truncation explicit in CAddrMan::Connected_()
CAddress::nTime is declared in src/protocol.h as uint32
but in CAddrMan::Connected_() it needs to be converted to int64
for subtraction from an int64 to avoid unsigned integer overflow,
and int64 values can be written to it with implicit integer truncation.

Fixes two UBSan errors reported by the addrman fuzzer

- a currently suppressed error
addrman.cpp:594:13: runtime error: unsigned integer overflow:
  513 - 100000000 cannot be represented in type 'unsigned int'
SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior addrman.cpp:594:13 in

- a non-suppressed error
addrman.cpp:595:22: runtime error: implicit conversion
  from type 'int64_t' (aka 'long') of value 68719478016 (64-bit, signed)
  to type 'uint32_t' (aka 'unsigned int') changed the value to 1280 (32-bit, unsigned)
SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior addrman.cpp:595:22

+++ src/addrman.cpp
@@ -592,8 +592,9 @@ void CAddrMan::Connected_(const CService& addr)
     // update info
     const int64_t nUpdateInterval{20 * 60};
     const int64_t now{GetAdjustedTime()};
-    if (now - info.nTime > nUpdateInterval) {
-        info.nTime = now;
+    // TODO: change CAddress::nTime from uint32 to int64 before the Year 2106.
+    if (now - static_cast<int64_t>(info.nTime) > nUpdateInterval) {
+        info.nTime = static_cast<uint32_t>(now);
     }
 }
